bump zod version
