+++ packages/api/package.json
@@ -23,11 +23,11 @@
     "@trpc/server": "^10.4.3",
     "discord.js": "^14.7.1",
     "superjson": "^1.9.1",
-    "zod": "^3.18.0"
+    "zod": "^3.20.6"
   },
   "devDependencies": {
-    "@answeroverflow/tsconfig": "*",
     "@answeroverflow/db-mock": "*",
+    "@answeroverflow/tsconfig": "*",
     "@vitest/coverage-c8": "^0.25.1",
     "dotenv-cli": "^6.0.0",
     "eslint": "^8.28.0",
