+++ packages/auth/package.json
@@ -15,8 +15,8 @@
   },
   "dependencies": {
     "@answeroverflow/db": "*",
-    "@answeroverflow/utils": "*",
     "@answeroverflow/tsconfig": "*",
+    "@answeroverflow/utils": "*",
     "@next-auth/prisma-adapter": "^1.0.5",
     "@prisma/client": "^4.7.1",
     "axios": "^1.2.2",
@@ -24,7 +24,7 @@
     "next-auth": "^4.17.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "zod": "^3.18.0"
+    "zod": "^3.20.6"
   },
   "devDependencies": {
     "@types/cookie": "^0.5.1",
