+++ packages/db-mock/package.json
@@ -18,7 +18,7 @@
     "@sapphire/bitfield": "^1.0.0",
     "discord-api-types": "^0.37.31",
     "type-fest": "^3.5.7",
-    "zod": "^3.20.2"
+    "zod": "^3.20.6"
   },
   "devDependencies": {
     "@answeroverflow/tsconfig": "*",
