+++ packages/db/package.json
@@ -16,13 +16,13 @@
     "test:watch": "yarn run with-env vitest"
   },
   "dependencies": {
+    "@answeroverflow/db-mock": "*",
     "@answeroverflow/elastic-types": "*",
     "@answeroverflow/prisma-types": "*",
-    "@answeroverflow/db-mock": "*",
     "@answeroverflow/utils": "*",
     "@sapphire/bitfield": "^1.0.0",
     "discord-api-types": "^0.37.30",
-    "zod": "^3.20.2"
+    "zod": "^3.20.6"
   },
   "devDependencies": {
     "@answeroverflow/tsconfig": "*",
