+++ packages/ui/package.json
@@ -30,7 +30,7 @@
     "next": "13.0.2",
     "recharts": "^2.3.2",
     "style-loader": "^3.3.1",
-    "zod": "^3.18.0"
+    "zod": "^3.20.6"
   },
   "devDependencies": {
     "@answeroverflow/tailwind-config": "*",
